fix: 정규식 수정, input component 분리, 폴더 구조 변경
