+++ src/pages/Auth/SignUpPage/signUpPage/entity.ts
@@ -1,0 +1,6 @@
+export interface SignUpFormData {
+  id: string;
+  email: string;
+  password: string;
+  passwordCheck: string;
+}
