+++ src/pages/Auth/SignUpPage/termsOfServicePage/TermsOfService.module.scss
